Properly handles transcodes with Duration:0
Some transcode formats in Subsonic don't provide valid Durations, which
results in SubFile.Stat with a Size of 0. This change at least makes some
sort of guessed filesize show up in Stat.
Additionally, when we receive the correct Size of the stream, we have
to update the size so that Fuse will pass along the entire data to the
application. However, since the ReadAll() function doesn't receive a shared
pointer to the SubFile, we can't update SubFile itself and need to track it
in a separate object.